InProgressChange: Working on Filter option
I tried several different ways to try and get the filter working, but couldn't quite get the syntax right.
I've commented out at the bottom of my js file if you can take a look and let me know if I was on the right track or not even in the ball park.